Feat: Add NN model for Iris dataset using mainly numpy
The model is created just using base python and numpy. sklearn is used to download the dataset and to divide it in training and test data. The goal is to create a deep learning model of a neural network using the most basic tools of python without any platform as tensorflow or pytorch to understand deeply all the parts in the architecture.